[#1474] Fix windows build

+++ iceoryx2-pal/posix/src/windows/mman.rs
@@ -19,7 +19,7 @@ use crate::posix::{
 };
 use crate::win32call;
 
-use iceoryx2_pal_concurrency_sync::atomic::AtomicU32;
+use iceoryx2_pal_concurrency_sync::atomic::AtomicU64;
 use iceoryx2_pal_concurrency_sync::cell::UnsafeCell;
 use iceoryx2_pal_concurrency_sync::strategy::mutex::Mutex;
 use iceoryx2_pal_concurrency_sync::WaitAction;
@@ -65,8 +65,8 @@ impl FileMappingsSet {
         self.mtx.lock(|atomic, value| {
             unsafe {
                 WaitOnAddress(
-                    (atomic as *const AtomicU32).cast(),
-                    (value as *const u32).cast(),
+                    (atomic as *const AtomicU64).cast(),
+                    (value as *const u64).cast(),
                     4,
                     INFINITE,
                 );
@@ -77,7 +77,7 @@ impl FileMappingsSet {
 
     fn unlock(&self) {
         self.mtx.unlock(|atomic| unsafe {
-            WakeByAddressSingle((atomic as *const AtomicU32).cast());
+            WakeByAddressSingle((atomic as *const AtomicU64).cast());
         });
     }
